requirements.txt
python-dotenv
openai== 1.26.0